requirements.txt
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
PySide2==5.14.1
python-dateutil==2.8.1
pytz==2019.3
randomcolor==0.4.4.5
shiboken2==5.14.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
